demo_code/pyproject.toml
[project]
name = "pdf-raptor-demo"
version = "0.1.0"
description = "Clean PDF processing with RAPTOR clustering using OpenAI or Gemini"
requires-python = ">=3.9,<3.13"
readme = "README.md"

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.5.1",
    "umap-learn>=0.5.6",
    "langchain>=0.2.5",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.5",
    "langchain-openai>=0.1.9",
    "langchain-google-genai>=1.0.0",
    "faiss-cpu>=1.8.0",
    "unstructured[all-docs]>=0.14.10",
    "pymupdf>=1.24.6",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.7.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.8.0",
    "sentence-transformers>=2.2.0",
    "huggingface-hub>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
